Support Parameterized JUnit tests
Create ParameterizedBehaviorTestFactory, a ParametersRunnerFactory creating Parameterized-compatible runners. Ideally this would just extend both BehaviorTestRunner *and* BlockJUnit4ClassRunnerWithParameters, but since that's not possible, we use the delegate pattern to simulate inheritance from a new SharedBehaviorTesting class, passing callbacks to access the superclass methods that were previously accessed directly in BehaviorTestRunner via **super**.
