+++ gradle.properties
@@ -8,7 +8,7 @@ hamcrest=org.hamcrest:hamcrest-all:1.3
 jacksonVersion=2.6.1
 javassist=org.javassist:javassist:3.19.0-GA
 jsr305=com.google.code.findbugs:jsr305:3.0.0
-junit=junit:junit:4.11
+junit=junit:junit:4.12
 mockito=org.mockito:mockito-core:1.10.8
 sorcerer=org.jvnet.sorcerer:sorcerer-javac:0.8
 truth=com.google.truth:truth:0.24
